Rearranged some code
