+++ go/lib/ctrl/seg/staticinfo_extn.go
@@ -1,9 +1,5 @@
 package seg
 
-import (
-	"math"
-)
-
 type LatencyInfo struct {
 	Egresslatency    uint16                  `capnp:"egressLatency"`
 	Intooutlatency   uint16                  `capnp:"ingressToEgressLatency"`
@@ -75,129 +71,4 @@ type StaticInfoExtn struct {
 	Bandwidth BandwidthInfo    `capnp:"bandwidth"`
 	Hops      InternalHopsInfo `capnp:"internalHops"`
 	Note      string           `capnp:"note"`
-}
-
-type Latintf struct {
-	Inter uint16            `json:"Inter"`
-	Intra map[uint16]uint16 `json:"Intra"`
-}
-
-type Bwintf struct {
-	Inter uint32            `json:"Inter"`
-	Intra map[uint16]uint32 `json:"Intra"`
-}
-
-type Geointf struct {
-	Longitude float32 `json:"Longitude"`
-	Latitude  float32 `json:"Latitude"`
-	Address   string  `json:"Address"`
-}
-
-type Hopintf struct {
-	Intra map[uint16]uint8 `json:"Intra"`
-}
-
-// Configdata is used to parse data from config.json.
-type Configdata struct {
-	Latency   map[uint16]Latintf `json:"Latency"`
-	Bandwidth map[uint16]Bwintf  `json:"Bandwidth"`
-	Linktype  map[uint16]string  `json:"Linktype"`
-	Geo       map[uint16]Geointf `json:"Geo"`
-	Hops      map[uint16]Hopintf `json:"Hops"`
-	Note      string             `json:"Note"`
-}
-
-// gatherlatency extracts latency values from a Configdata struct and
-// inserts them into the LatencyInfo portion of a StaticInfoExtn struct.
-func (latinf *LatencyInfo) Gatherlatency(cfgdata Configdata, peers map[uint16]bool, egifID uint16, inifID uint16) {
-	latinf.Egresslatency = cfgdata.Latency[egifID].Inter
-	latinf.Intooutlatency = cfgdata.Latency[egifID].Intra[inifID]
-	for subintfid, intfdelay := range cfgdata.Latency[egifID].Intra {
-		if !(peers[subintfid]) {
-			if subintfid > egifID {
-				var latpair Latencychildpair
-				latpair.Intradelay = intfdelay
-				latpair.Interface = subintfid
-				latinf.Childlatencies = append(latinf.Childlatencies, latpair)
-			}
-		} else {
-			var lattriple Latencypeeringtriplet
-			lattriple.IntfID = subintfid
-			lattriple.Interdelay = cfgdata.Latency[subintfid].Inter
-			lattriple.IntraDelay = intfdelay
-			latinf.Peeringlatencies = append(latinf.Peeringlatencies, lattriple)
-		}
-	}
-}
-
-// gatherbw extracts bandwidth values from a Configdata struct and
-// inserts them into the BandwidthInfo portion of a StaticInfoExtn struct.
-func (bwinf *BandwidthInfo) Gatherbw(cfgdata Configdata, peers map[uint16]bool, egifID uint16, inifID uint16) {
-	bwinf.EgressBW = cfgdata.Bandwidth[egifID].Inter
-	bwinf.IntooutBW = cfgdata.Bandwidth[egifID].Intra[inifID]
-	for subintfid, intfbw := range cfgdata.Bandwidth[egifID].Intra {
-		var minbw uint32
-		if subintfid > egifID {
-			if peers[subintfid] {
-				minbw = uint32(math.Min(float64(intfbw), float64(cfgdata.Bandwidth[subintfid].Inter)))
-			} else {
-				minbw = intfbw
-			}
-			var bwpair BWPair
-			bwpair.BW = minbw
-			bwpair.IntfID = subintfid
-			bwinf.BWPairs = append(bwinf.BWPairs, bwpair)
-		}
-	}
-}
-
-// gatherlinktype extracts linktype values from a Configdata struct and
-// inserts them into the LinktypeInfo portion of a StaticInfoExtn struct.
-func (ltinf *LinktypeInfo) Gatherlinktype(cfgdata Configdata, peers map[uint16]bool, egifID uint16) {
-	ltinf.EgressLT = cfgdata.Linktype[egifID]
-	for intfid, intfLT := range cfgdata.Linktype {
-		if peers[intfid] {
-			var ltpair LTPeeringpair
-			ltpair.IntfLT = intfLT
-			ltpair.IntfID = intfid
-			ltinf.Peeringlinks = append(ltinf.Peeringlinks, ltpair)
-		}
-	}
-}
-
-// gatherhops extracts hop values from a Configdata struct and
-// inserts them into the InternalHopsinfo portion of a StaticInfoExtn struct.
-func (nhinf *InternalHopsInfo) Gatherhops(cfgdata Configdata, egifID uint16, inifID uint16) {
-	nhinf.Intououthops = cfgdata.Hops[egifID].Intra[inifID]
-	for subintfid, hops := range cfgdata.Hops[egifID].Intra {
-		if subintfid > egifID {
-			var hoppair Hoppair
-			hoppair.Hops = hops
-			hoppair.IntfID = subintfid
-			nhinf.Hoppairs = append(nhinf.Hoppairs, hoppair)
-		}
-	}
-}
-
-// gathergeo extracts geo values from a Configdata struct and
-// inserts them into the GeoInfo portion of a StaticInfoExtn struct.
-func (geoinf *GeoInfo) Gathergeo(cfgdata Configdata) {
-	for intfid, loc := range cfgdata.Geo {
-		var assigned = false
-		for k := 0; k < len(geoinf.Locations); k++ {
-			if (loc.Longitude == geoinf.Locations[k].GPSData.Longitude) && (loc.Latitude == geoinf.Locations[k].GPSData.Latitude) && (loc.Address == geoinf.Locations[k].GPSData.Address) && (!assigned) {
-				geoinf.Locations[k].IntfIDs = append(geoinf.Locations[k].IntfIDs, intfid)
-				assigned = true
-			}
-		}
-		if !assigned {
-			var locaction Location
-			locaction.GPSData.Longitude = loc.Longitude
-			locaction.GPSData.Latitude = loc.Latitude
-			locaction.GPSData.Address = loc.Address
-			locaction.IntfIDs = append(locaction.IntfIDs, intfid)
-			geoinf.Locations = append(geoinf.Locations, locaction)
-			assigned = true
-		}
-	}
 }
